--- a/test/Riok.Mapperly.Tests/Descriptors/MemberPathCandidateBuilderTest.cs
+++ b/test/Riok.Mapperly.Tests/Descriptors/MemberPathCandidateBuilderTest.cs
@@ -5,9 +5,28 @@ namespace Riok.Mapperly.Tests.Descriptors;
 public class MemberPathCandidateBuilderTest
 {
     [Theory]
+    [InlineData("", new string[] { })]
+    [InlineData("a", new[] { "a" })]
+    [InlineData("A", new[] { "A" })]
+    [InlineData("aA", new[] { "aA", "a.A" })]
+    [InlineData("AB", new[] { "AB", "A.B" })]
     [InlineData("Value", new[] { "Value" })]
     [InlineData("MyValue", new[] { "MyValue", "My.Value" })]
     [InlineData("MyValueId", new[] { "MyValueId", "My.ValueId", "MyValue.Id", "My.Value.Id" })]
+    [InlineData(
+        "MyValueIdNum",
+        new[]
+        {
+            "MyValueIdNum",
+            "My.ValueIdNum",
+            "MyValue.IdNum",
+            "My.Value.IdNum",
+            "MyValueId.Num",
+            "My.ValueId.Num",
+            "MyValue.Id.Num",
+            "My.Value.Id.Num"
+        }
+    )]
     public void BuildMemberPathCandidatesShouldWork(string name, string[] chunks)
     {
         MemberPathCandidateBuilder