… src/Riok.Mapperly/Helpers/StringChunker.cs
… test/Riok.Mapperly.Tests/Helpers/StringChunkerTest.cs