## Description
Optimized `BuildMemberPathCandidates` by using a faster string concat and early yield.

- Changed `BuildMemberPathCandidates` to immediately `yield return` the full string, this way the same string isn't cloned. This is significant because most matches are identical.
- `BuildName` uses an array of lengths of the split members to get the corresponding substring of the permutation. In this case `Substring` is faster than `string.Concat` because .NET already knows the final length of the string so a simple copy is performed. `string.Concat` doesn't know the size of the final string and has to iterate through the strings, adding each one to the `ValueStringBuilder`.
- Added more `BuildMemberPathCandidatesShouldWork` and adapted `StringChunkerTest`.
- Deleted `Chunk` and all corresponding tests.

See #479, this change doesn't alter algorithm, so the nasty big O behaviour is still present, but it will reduce memory and yield a small speed boost.

## Benchmarks
Saved 300KB of memory for larger compilations. Note that for large unmatching members with capital letters, we see a speed up of 33% and almost half the memory usage.

**Old**
|       Method |      Mean |     Error |    StdDev |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|----------:|----------:|---------:|-----------:|
|      Compile |  1.614 ms | 0.0232 ms | 0.0206 ms |   74.2188 |  15.6250 |  167.84 KB |
| LargeCompile | 59.971 ms | 0.9963 ms | 1.1473 ms | 1428.5714 | 285.7143 | 9283.43 KB |

|       Method |      Mean |     Error |    StdDev |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|----------:|----------:|---------:|-----------:|
|      Compile |  1.605 ms | 0.0301 ms | 0.0267 ms |   78.1250 |  19.5313 |  170.23 KB |
| LargeCompile | 59.858 ms | 1.1878 ms | 2.1114 ms | 1428.5714 | 285.7143 | 9268.93 KB |

|       Method |      Mean |     Error |    StdDev |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|----------:|----------:|---------:|-----------:|
|      Compile |  1.621 ms | 0.0213 ms | 0.0188 ms |   82.0313 |  19.5313 |  167.99 KB |
| LargeCompile | 60.269 ms | 0.7791 ms | 0.8660 ms | 1428.5714 | 428.5714 | 9268.14 KB |

**New**
|       Method |      Mean |     Error |    StdDev |    Median |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|----------:|----------:|----------:|---------:|-----------:|
|      Compile |  1.685 ms | 0.0502 ms | 0.1424 ms |  1.625 ms |   82.0313 |  19.5313 |  165.29 KB |
| LargeCompile | 60.405 ms | 1.8495 ms | 5.1249 ms | 57.832 ms | 1428.5714 | 428.5714 | 8960.02 KB |

|       Method |      Mean |     Error |    StdDev |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|----------:|----------:|---------:|-----------:|
|      Compile |  1.530 ms | 0.0234 ms | 0.0207 ms |   78.1250 |  15.6250 |  166.27 KB |
| LargeCompile | 56.023 ms | 0.5765 ms | 0.5393 ms | 1375.0000 | 250.0000 | 8961.08 KB |

|       Method |      Mean |     Error |    StdDev |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|----------:|----------:|---------:|-----------:|
|      Compile |  1.567 ms | 0.0240 ms | 0.0388 ms |   78.1250 |  15.6250 |  166.45 KB |
| LargeCompile | 58.106 ms | 1.0849 ms | 1.9563 ms | 1375.0000 | 375.0000 | 8967.22 KB |

**Old with long unmatching full caps member**
|  Method |     Mean |    Error |   StdDev |      Gen0 |    Gen1 | Allocated |
|-------- |---------:|---------:|---------:|----------:|--------:|----------:|
| Compile | 23.73 ms | 0.250 ms | 0.209 ms | 6406.2500 | 62.5000 |    9.6 MB |

**New with long unmatching full caps member**
|  Method |     Mean |    Error |   StdDev |      Gen0 |    Gen1 | Allocated |
|-------- |---------:|---------:|---------:|----------:|--------:|----------:|
| Compile | 17.57 ms | 0.314 ms | 0.590 ms | 3656.2500 | 62.5000 |   5.52 MB |

### Substring vs string.Concat
Skip 2 Take 2
|    Method |      Mean |    Error |   StdDev |    Median |   Gen0 | Allocated |
|---------- |----------:|---------:|---------:|----------:|-------:|----------:|
| SubString |  30.87 ns | 2.656 ns | 7.833 ns |  28.45 ns | 0.0306 |      48 B |
|    Concat | 166.60 ns | 2.966 ns | 3.750 ns | 167.23 ns | 0.0918 |     144 B |

Take first
|    Method |     Mean |    Error |   StdDev |   Gen0 | Allocated |
|---------- |---------:|---------:|---------:|-------:|----------:|
| SubString | 18.81 ns | 0.459 ns | 0.659 ns | 0.0204 |      32 B |
|    Concat | 73.55 ns | 1.366 ns | 1.278 ns | 0.0305 |      48 B |