@TimothyMakkison 2023-06-06
> Build a list of all upper case char indices in the string, loop through all permutations and build the string based on the indices. I did not benchmark it yet, but I think this could be an easier to understand solution with an acceptable performance. WDYT?

This is very similar to what I originally wanted. It's why my variable names are a little wrong (offsets not indices) 😂 
I ended up using the offsets because I kept getting a nasty OBOE and looping infintely. I'll use your offset system.