@TimothyMakkison 2023-06-06
- Deleted `StringChunker` and associated tests.
- Changed `GetPascalCaseSplitIndices` to use a for loop instead of foreach. 
- Modifed `splitIndices` in `BuildName` to accept an array instead of `IEnumerable`.

### Benchmarks
Doesn't appear to be a regression.

#### New
|       Method |      Mean |     Error |     StdDev |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|-----------:|----------:|---------:|-----------:|
|      Compile |  2.317 ms | 0.1589 ms |  0.4686 ms |   78.1250 |  15.6250 |  166.27 KB |
| LargeCompile | 78.160 ms | 3.8618 ms | 11.3260 ms | 1400.0000 | 200.0000 | 8964.42 KB |

|       Method |      Mean |     Error |     StdDev |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|-----------:|----------:|---------:|-----------:|
|      Compile |  1.997 ms | 0.0979 ms |  0.2856 ms |   78.1250 |  15.6250 |  167.88 KB |
| LargeCompile | 78.429 ms | 5.0801 ms | 14.6573 ms | 1250.0000 | 250.0000 | 8947.21 KB |

#### Unmodifed
|       Method |      Mean |     Error |     StdDev |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|-----------:|----------:|---------:|-----------:|
|      Compile |  2.063 ms | 0.0981 ms |  0.2829 ms |   78.1250 |  15.6250 |  169.48 KB |
| LargeCompile | 78.972 ms | 3.5745 ms | 10.3704 ms | 1400.0000 | 400.0000 | 9266.61 KB |

|       Method |      Mean |     Error |     StdDev |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|-----------:|----------:|---------:|-----------:|
|      Compile |  2.778 ms | 0.2019 ms |  0.5952 ms |   93.7500 |  23.4375 |  169.26 KB |
| LargeCompile | 80.579 ms | 4.3258 ms | 11.7687 ms | 1500.0000 | 500.0000 | 9247.04 KB |


#### New with long member
|       Method |      Mean |     Error |    StdDev |       Gen0 |      Gen1 | Allocated |
|------------- |----------:|----------:|----------:|-----------:|----------:|----------:|
|      Compile |  25.62 ms |  1.649 ms |  4.837 ms |  3642.8571 |   71.4286 |   5.55 MB |