@latonz 2023-06-07
Mapperly doesn't reference `System.Memory`. The .net framework integration tests should reference net48 🤔  Therefore I'm wondering why Span does work in the integration tests too...